# Bekbook/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Mahalla Kutubxona API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transactions:
    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Create a lending transaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getMyTransactions
      tags: [transactions]
      summary: Get my lending transactions
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, returned, overdue]
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionsListResponse"
...
  /invoices:
    post:
      operationId: createInvoice
      tags: [invoices]
      summary: Create a new nakladnoy (kirim or chiqim)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInvoiceBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /invoices/store/{storeId}:
    get:
      operationId: getStoreInvoices
      tags: [invoices]
      summary: Get all nakladnoylar for a store
      security:
        - bearerAuth: []
      parameters:
        - name: storeId
          in: path
          required: true
          schema:
            type: integer
        - name: type
          in: query
          schema:
            type: string
            enum: [kirim, chiqim]
      responses:
        "200":
          description: Invoices list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvoicesListResponse"
  /invoices/{invoiceId}:
    get:
      operationId: getInvoice
      tags: [invoices]
      summary: Get a single invoice with items
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Invoice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invoice"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteInvoice
      tags: [invoices]
      summary: Delete an invoice
      security:
        - bearerAuth: []
      parameters:
        - name: invoiceId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    UserSearchResult:
      type: object
      properties:
        id:
          type: integer
        readerId:
          type: string
          nullable: true
        name:
          type: string
        phone:
          type: string
          nullable: true
        avatar:
          type: string
          nullable: true
      required:
        - id
        - name
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        lenderId:
          type: integer
        bookId:
          type: integer
          nullable: true
        storeBookId:
          type: integer
          nullable: true
        borrowerUserId:
          type: integer
          nullable: true
        borrowerName:
          type: string
        borrowerPhone:
          type: string
          nullable: true
        issuedAt:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnedAt:
          type: string
          format: date-time
          nullable: true
        status:
          type: string
          enum: [active, returned, overdue]
        finePerDay:
          type: number
        fineAmount:
          type: number
        notes:
          type: string
          nullable: true
        bookTitle:
          type: string
          nullable: true
        borrowerConfirmedAt:
          type: string
          format: date-time
          nullable: true
        returnConfirmedAt:
          type: string
          format: date-time
          nullable: true
        borrowerUser:
          $ref: "#/components/schemas/UserSearchResult"
          nullable: true
      required:
        - id
        - lenderId
        - borrowerName
        - issuedAt
        - dueDate
        - status
        - finePerDay
        - fineAmount
    TransactionsListResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
      required:
        - transactions
        - total
...
    InvoiceItem:
      type: object
      properties:
        id:
          type: integer
        invoiceId:
          type: integer
        storeBookId:
          type: integer
          nullable: true
        title:
          type: string
        author:
          type: string
          nullable: true
        isbn:
          type: string
          nullable: true
        quantity:
          type: integer
        unitPrice:
          type: number
        reason:
          type: string
          nullable: true
        inventoryNumbers:
          type: string
          nullable: true
      required: [id, invoiceId, title, quantity, unitPrice]
    Invoice:
      type: object
      properties:
        id:
          type: integer
        storeId:
          type: integer
        type:
          type: string
          enum: [kirim, chiqim]
        number:
          type: string
        date:
          type: string
        supplier:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdBy:
          type: integer
          nullable: true
        createdAt:
          type: string
          format: date-time
        items:
          type: array
          items:
            $ref: "#/components/schemas/InvoiceItem"
      required: [id, storeId, type, number, date, createdAt]
    InvoicesListResponse:
      type: object
      properties:
        invoices:
          type: array
          items:
            $ref: "#/components/schemas/Invoice"
      required: [invoices]
    CreateInvoiceBody:
      type: object
      properties:
        type:
          type: string
          enum: [kirim, chiqim]
        number:
          type: string
        date:
          type: string
        supplier:
          type: string
        notes:
          type: string
        items:
          type: array
          items:
            type: object
            properties:
              storeBookId:
                type: integer
              title:
                type: string
              author:
                type: string
              isbn:
                type: string
              quantity:
                type: integer
              unitPrice:
                type: number
              reason:
                type: string
              inventoryNumbers:
                type: string
            required: [title, quantity]
      required: [type, number, date, items]
...
    CreateTransactionBody:
      type: object
      properties:
        bookId:
          type: integer
        storeBookId:
          type: integer
        borrowerUserId:
          type: integer
          description: Registered user ID of the borrower
        borrowerName:
          type: string
        borrowerPhone:
          type: string
        dueDate:
          type: string
          format: date-time
        finePerDay:
          type: number
          description: Fine per overdue day in so'm
        notes:
          type: string
      required:
        - borrowerName
        - dueDate